+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add configurable timeouts for LFS uploads
 
+### Changed
+- Decouple library from OS environment variables
+
 ### Fixed
 - Prevent URL-encoded slashes in repository IDs